+++ NEWS.md
@@ -1,5 +1,10 @@
 # connectapi (development version)
 
+## New features
+
+- New `get_usage()` function returns content usage data from Connect's `GET
+  v1/instrumentation/content/hits` endpoint on Connect v2025.04.0 and higher.
+  (#390)
 
 ## Enhancements and fixes